Split service
